+++ ReleaseNotes.md
@@ -3,3 +3,4 @@ Here's a list of the latest changes:
 **3.18.0**
 
 - [Change - Guests will be sent the ticket now](https://trello.com/c/r7biHc3s/584-each-guest-should-receive-the-ticket-along-with-an-invite)
+- [Enhancement - Guests will be sent a different notification explaining that changes were made to their ticket or event.](https://trello.com/c/CHgDUYTW/585-resending-ticket-notification)